+++ packages/toprf-secure-backup/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.8.0]
 
-### Uncategorized
+### Changed
 
 - chore: dep updates ([#143](https://github.com/MetaMask/toprf-secure-backup/pull/143))