+++ packages/toprf-secure-backup/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - chore: dep updates ([#143](https://github.com/MetaMask/toprf-secure-backup/pull/143))
+- Updated: @metamask/create-release-branch ([#144](https://github.com/MetaMask/toprf-secure-backup/pull/144))
 
 ## [0.7.1]